test(onboard): split resume recreation coverage
